backend/requirements.txt
fastapi[standard]
sqlalchemy
pydantic-settings
python-dotenv
python-multipart
alembic
supabase
httpx
pytest